:paw_prints: Posts now support multiple images
## Timeline
Thanks to a very helpful YouTube tutorial, I was able to mimic Instagram's image slideshow thing.
- Posts can have as many images as we want (in theory) and as few as none
- Need to add a pin button to display the location of the animal sighting on the map when pressed. (For now we can just have this open in the default maps app)
- Pictures have to be square. No I will not put the effort required into fixing this.
- Need to make sure my renderItem functions follow React performance best practices. Don't know what those are.

## Navigation
I do think it's important we use flex here instead of padding in App.js, because I believe that is device dependent, where flex is not.
On an iOS device I feel the extra space is justified because the home bar is in the way on the bottom, so now it changes the size based on Platform.