:art: Added Themes
Added themes and then implemented them. This could be very nice for if we decide to change major color schemes often, but I don't know why we would do that. I saw someone using it in a tutorial and  thought it was cool :shrug:

Appearance.getColorScheme() should return if the user's phone is in light or dark theme. This way we can respond to the user's preferences :+1: