:camera: Updated Icons
Updated Icons to the version with the hand properly covering the magnifying glass

In PostPage.js I added a wrapper that will minimize the keyboard when tapping the background.
I also changed the font color of the submit button because the background of the pressable changed with theme (since it's transparent) but the text did not

In TimeLine.js I added an invisible button on top of the header that scrolls to the top of the timeline.