:recycle: Refactored the database code
I made a component (dataStorage.js) that handles the initialization of the Firebase stuff. This way it shouldn't be re-rendered on every page and won't have to be passed through the navigation parameters

in the post page I also added {Description: text} to the upload

I left all the commented out code in case someone wants it for reference

The TimeLine page now pulls posts from the "FakeStrays" collection. I'm doing this because making the format fit does not require changing the entirety of the posting page

*Note: * I did delete the documents from the firestore database. We weren't using them and I found it difficult to find the pushes I was making so I thought this was fine.