:bug: Fixed the issue with updating unmounted component
*Probably*
When logging in, sometimes the user would encounter an error where something would update after the LogIn screen had been replaced. I had suspicions this was the Modal, so I replaced it with react-native's built in modal instead of the community one.

While replacing it I noticed we reuse some of the same code for a loading modal and made it it's own component